python-client/requirements.txt
typesense==0.16.0
python-dotenv